# docs/openapi.yaml
openapi: 3.0.3
info:
  title: One Piece TCG Shelf API
  version: 1.0.0
  description: Public read APIs and authenticated trade offer create.
servers:
  - url: /
paths:
  /api/v1/profiles/{username}:
    get:
      summary: Public profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile
        "404":
          description: Not found
  /api/v1/profiles/{username}/collection:
    get:
      summary: Public collection (no valuations)
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection cards
  /api/v1/profiles/{username}/showcase:
    get:
      summary: Public showcase slots
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Showcase
  /api/v1/profiles/{username}/activity:
    get:
      summary: Public activity feed
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity events
  /api/v1/trade-offers:
    post:
      summary: Create trade offer (authenticated)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_collection_id]
              properties:
                target_collection_id:
                  type: string
                  format: uuid
                message:
                  type: string
      responses:
        "201":
          description: Offer created
        "401":
          description: Unauthorized
        "400":
          description: Validation / business rule error
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase Auth session cookies set by the app after login.